+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- Dart 3 Support ([#1220](https://github.com/getsentry/sentry-dart/pull/1220))
+
 ### Breaking Changes
 
 - Enable APM features by default ([#1217](https://github.com/getsentry/sentry-dart/pull/1217))